+++ Finance/Business/SHGEquityInvestSendViewController.m
@@ -208,7 +208,7 @@
     }
     NSInteger leftTextFieldString = 10000 * [self.leftMoneyTextField.text  integerValue];
     NSInteger rightTextFieldString = [self.monenyTextField.text integerValue];
-    NSString *money = [NSString stringWithFormat:@"%ld",leftTextFieldString + rightTextFieldString];
+    NSString *money = [NSString stringWithFormat:@"%ld", (long)(leftTextFieldString + rightTextFieldString)];
     if (self.leftMoneyTextField.text.length == 0 && self.monenyTextField.text.length == 0) {
         money = @"";
     }
@@ -548,20 +548,9 @@
         NSMutableArray * array = [[SHGGloble sharedGloble] editBusinessKeysForValues:nameArray middleContentArray:resultArray];
         NSLog(@"%@", array);
         //金额
-//        self.phoneNumTextField.text = [array objectAtIndex:4];
-//        
-//        NSString *money = [array objectAtIndex:1];
-//        NSString *str = [money substringWithRange:NSMakeRange(money.length - 1, 1)];
-//        if ([money isEqualToString:@"暂未说明"]) {
-//            self.monenyTextField.text = @"";
-//        } else if ([str isEqualToString:@"亿"]){
-//            self.monenyTextField.text = [[money substringToIndex:money.length - 1] stringByAppendingString:@"0000"];
-//        } else{
-//            self.monenyTextField.text = [money substringWithRange:NSMakeRange(0, money.length - 1)];
-//        }
         NSInteger money = [self.object.investmoney integerValue];
-        NSString *leftText= [NSString stringWithFormat:@"%ld",money/10000];
-        NSString *rightText= [NSString stringWithFormat:@"%ld",money%10000];
+        NSString *leftText = [NSString stringWithFormat:@"%ld", (long)money/10000];
+        NSString *rightText = [NSString stringWithFormat:@"%ld", (long)money%10000];
         if (![leftText isEqualToString:@"0"]) {
             self.leftMoneyTextField.text = leftText;
         } else{
